Prevent jit-compilation from loading classes in jit-on-first-use
Unlike with normal jit, jit-on-first-use (-Xjitthreshold:0) causes the
jit-compilation to be done synchronously on the current thread. Since
this thread is not a runtime thread it will allow the jit to load
classes even on a debuggable runtime. This can cause some tests to
fail. To fix this (and prevent any future issues) we mark threads
running in jit-code for jit-on-first-use as runtime-threads
temporarily.

Test: ./test/run-test --host --jit --runtime-option -Xjitthreshold:0 1953
Change-Id: I06410bc3cbdb28ba51e907b71556481ba405c1d0