Add test where we test jit & redefinition suitable for CTS
We didn't have any tests of redefinition and JIT interaction suitable
for CTS since most of them used internal runtime-functions to force
functions to be jitted. This test tries to ensure a particular method
is jitted by running it 20000 times and waiting a bit.

This should give us confidence that the jit is not inlining methods
which would prevent us from redefining classes.

Test: ./test.py --host
Bug: 119063276
Change-Id: Ibde0d19ef0893ece3a494a547cff384a1248a21f