Configure the full set of libraries for switching back from platform to runtime (aka default) namespaces.
In CTS dalvikvm loads libandroid.so which is only in /system/lib, and then
that loads libnativehelper.so which should be loaded from the APEX again.

It's not certain this list needs to be a full copy of the list of
namespace.default.link.runtime.shared_libs in
system/core/rootdir/etc/ld.config.txt, but it's simpler that way.

Bug: 121384464
Bug: 119867084
Bug: 113373927
Test: Flash and reboot device
Test: atest --test-mapping cts/tools/vm-tests-tf/TEST_MAPPING
Test: cts-tradefed run commandAndExit cts-dev -m CtsJniTestCases
Test: cts-tradefed run commandAndExit cts-dev -m CtsCompilationTestCases
Change-Id: Iecd27b8c9d4b7b80141f8a7f9eb6f062f3885206